+++ bot/draw.py
